+++ src/validator.py
@@ -35,7 +35,7 @@ class Validator:
             for data, target in self.test_loader:
                 data, target = data.to(self.device), target.to(self.device)
                 outputs = model(data)
-                loss = criterion(outputs, target)
+                loss = criterion(outputs, target.type(torch.int64))
                 total_loss += loss.item()
                 _, predicted = torch.max(outputs.data, 1)
                 total += target.size(0)
